requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.7
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pytest==8.3.2
python-dotenv==1.0.1
requests==2.32.3
requests-mock==1.12.1
urllib3==2.2.2